+++ app/assets/javascripts/ci_api/controllers/ProactiveChatController.js
@@ -1,0 +1,13 @@
+
+import CommonChatController from './CommonChatController'
+
+export default class ProactiveChatController {
+    constructor() {
+        this.sdk = null;
+        this.commonChatController = new CommonChatController();
+    }
+
+    launchProactiveChat() {
+        this.commonChatController._launchChat();
+    }
+}
