+++ app/assets/javascripts/ci_api/controllers/ReactiveChatController.js
@@ -1,0 +1,35 @@
+import ClickToChatButtons from '../ClickToChatButtons'
+import ClickToChatButton from '../ClickToChatButton'
+import CommonChatController from './CommonChatController'
+import * as DisplayState from '../NuanceDisplayState'
+
+const c2cDisplayStateMessages = {
+    [DisplayState.OutOfHours]: "Out of hours",
+    [DisplayState.Ready]: "Ask HMRC a question",
+    [DisplayState.Busy]: "All advisers are busy",
+    [DisplayState.ChatActive]: "In progress"
+};
+
+export default class ReactiveChatController {
+    constructor() {
+        this.sdk = null;
+        this.c2cButtons = new ClickToChatButtons((c2cIdx) => this._onC2CButtonClicked(c2cIdx), c2cDisplayStateMessages);
+        this.commonChatController = new CommonChatController();
+    }
+
+    addC2CButton(c2cObj, divID, buttonClass) {
+        this.c2cButtons.addButton(
+            c2cObj,
+            new ClickToChatButton(document.getElementById(divID), buttonClass)
+        );
+    }
+
+    _onC2CButtonClicked(c2cIdx) {
+        this.sdk.onC2CClicked(c2cIdx, (state) => {
+            console.log("onC2CClicked callback:");
+            console.log(state);
+            this.commonChatController._launchChat();
+        });
+    }
+
+}
