+++ app/assets/javascripts/ci_api/views/popup/PopupContainerHtml.js
@@ -1,4 +1,5 @@
-export const PopupContainerHtml = `
+'use strict';
+export const ContainerHtml = `
 <div id="ciapiSkinContainer">
     <div id="ciapiSkinHeader">
         <div id="ciapiTitleBarLogo"><img src='/ask-hmrc/assets/media/logo.png' alt=""></div>
@@ -45,46 +46,5 @@ export const PopupContainerHtml = `
         <div id="logo-white"><img src="/ask-hmrc/assets/media/logo-white.png" alt="" aria-hidden="true"></div>
         <h2 class="govuk-heading-s">Ask HMRC a Question</h2>
     </button>
-</div>
-`
-
-export const EmbeddedContainerHtml = `
-<div id="ciapiSkinContainer">
-    <div id="ciapiSkinHeader">
-        <div id="print">
-            <button class="govuk-button govuk-button--secondary" data-module="govuk-button">
-                Print or save
-            </button>
-        </div>
-        <div id="sound">
-            <button class="govuk-button govuk-button--secondary" data-module="govuk-button">
-                Turn sound on
-            </button>
-        </div>
-    </div>
-    <div id="ciapiChatComponents">
-        <div id="ciapiSkinChatTranscript" role="log" tabindex="0">
-            <div id="skipToBottom"><a id="skipToBottomLink" href="#skipToTopLink" class="govuk-skip-link">Skip to bottom of conversation</a></div>
-            <p class="info"><img src="/ask-hmrc/assets/media/intro-warn.svg" alt="">You are currently chatting with a computer.</p>
-        </div>
-        <div id="ciapiSkinFooter">
-            <div>
-                <div id="ciapiInput"><textarea
-                    id="custMsg"
-                    aria-label="Type your message here"
-                    placeholder="Type your message here"
-                    class="govuk-textarea"
-                    cols="50"
-                    name="comments"></textarea>
-                </div>
-                <div id="ciapiSend">
-                    <button id="ciapiSkinSendButton" class="govuk-button">Send</button>
-                </div>
-            </div>
-            <div id="ciapiClose">
-                <button id="ciapiSkinCloseButton" style="margin-top:-20px;" class="govuk-button govuk-button--secondary" data-module="govuk-button">End chat</button>
-            </div>
-        </div>
-    </div>
 </div>
 `
